+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added iteration support to `Pipeline`, allowing pipeline steps to be iterated in their configured order.
 - Added `Pipeline.result` property to retrieve the most recent pipeline result.
 - Added `Pipeline.error` property to re-raise the most recent exception raised by a pipeline step.
+- Added `run_now` and `run_args` parameters to `Pipeline` to support automatically starting the pipeline with custom `run()` arguments during initialization.
 
 ## 0.2.4
 - Documentation-only release: Fix heading level in README.